build with g++

+++ include/sys/ioctl.h
@@ -105,13 +105,13 @@ of the xfer buffer in CX. Aaaaargh
 #define IOC_VOID        0x20000000      /* no parameters */
 #define IOC_OUT         0x40000000      /* copy out parameters */
 #define IOC_IN          0x80000000      /* copy in parameters */
-#define IOC_INOUT       (IOC_IN|IOC_OUT)
+#define IOC_INOUT       (int)(IOC_IN|IOC_OUT)
 /* the 0x20000000 is so we can distinguish new ioctl's from old */
-#define _IO(x,y)        (IOC_VOID|(x<<8)|y)
-#define _IOR(x,y,t)     (IOC_OUT|((sizeof(t)&IOCPARM_MASK)<<16)|(x<<8)|y)
-#define _IOW(x,y,t)     (IOC_IN|((sizeof(t)&IOCPARM_MASK)<<16)|(x<<8)|y)
+#define _IO(x,y)        (int)(IOC_VOID|(x<<8)|y)
+#define _IOR(x,y,t)     (int)(IOC_OUT|((sizeof(t)&IOCPARM_MASK)<<16)|(x<<8)|y)
+#define _IOW(x,y,t)     (int)(IOC_IN|((sizeof(t)&IOCPARM_MASK)<<16)|(x<<8)|y)
 /* this should be _IORW, but stdio got there first */
-#define _IOWR(x,y,t)    (IOC_INOUT|((sizeof(t)&IOCPARM_MASK)<<16)|(x<<8)|y)
+#define _IOWR(x,y,t)    (int)(IOC_INOUT|((sizeof(t)&IOCPARM_MASK)<<16)|(x<<8)|y)
 #endif /* _IO */
 
 /*
